satellite-processor/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pystac-client==0.8.3
stackstac==0.5.1
numpy==1.26.4
Pillow==11.1.0
rasterio==1.3.11
shapely==2.0.6
pyproj==3.7.0
httpx==0.28.1
python-dotenv==1.0.1
